+++ Jellyfin.Plugin.JellyfinHelper/PluginPages/js/Overview.js
@@ -92,14 +92,18 @@ function fillOverviewData(data) {
 
     var overviewHtml = '';
     overviewHtml += '<div class="stats-grid">';
-    overviewHtml += '<div class="stat-card"><h3>' + mi('movie') + escHtml(T('movieVideoData', 'Video Data - Movies')) + '</h3>';
+    overviewHtml += '<div class="stat-card stat-card-link" role="button" tabindex="0" data-codec-explore-library="'
+        + CODEC_EXPLORER_TYPE_MOVIES + '" title="' + escAttr(T('explorerOpenTooltip', 'Open in Library Explorer')) + '">';
+    overviewHtml += '<h3>' + mi('movie') + escHtml(T('movieVideoData', 'Video Data - Movies')) + '</h3>';
     overviewHtml += '<p class="stat-value">' + formatBytes(data.TotalMovieVideoSize) + '</p>';
     var movieFiles = 0;
     for (const movie of movies) movieFiles += movie.VideoFileCount;
     overviewHtml += '<p class="stat-detail">' + movieFiles + ' ' + (movieFiles === 1 ? escHtml(T('file', 'file')) : escHtml(T('files', 'files'))) + ' ' + escHtml(T('across', 'across')) + ' ' + movies.length + ' ' + escHtml(T('libraries', 'libraries')) + '</p>';
     overviewHtml += '</div>';
 
-    overviewHtml += '<div class="stat-card"><h3>' + mi('tv') + escHtml(T('tvVideoData', 'Video Data - TV Shows')) + '</h3>';
+    overviewHtml += '<div class="stat-card stat-card-link" role="button" tabindex="0" data-codec-explore-library="'
+        + CODEC_EXPLORER_TYPE_TVSHOWS + '" title="' + escAttr(T('explorerOpenTooltip', 'Open in Library Explorer')) + '">';
+    overviewHtml += '<h3>' + mi('tv') + escHtml(T('tvVideoData', 'Video Data - TV Shows')) + '</h3>';
     overviewHtml += '<p class="stat-value">' + formatBytes(data.TotalTvShowVideoSize) + '</p>';
     var tvFiles = 0;
     for (const show of tvShows) tvFiles += show.VideoFileCount;
@@ -156,9 +160,9 @@ function fillOverviewData(data) {
 
     for (const lib of libraries) {
         overviewHtml += '<tr>';
-        overviewHtml += '<td>' + escHtml(lib.LibraryName)
-            + '<button class="codec-explore-link" data-codec-explore-library="' + escAttr(lib.LibraryName) + '">'
-            + escHtml(T('exploreInCodecs', 'Explore')) + '</button></td>';
+        overviewHtml += '<td><button class="codec-explore-link" data-codec-explore-library="' + escAttr(lib.LibraryName) + '"'
+            + ' title="' + escAttr(T('explorerOpenTooltip', 'Open in Library Explorer')) + '">'
+            + escHtml(lib.LibraryName) + '</button></td>';
         overviewHtml += '<td>' + getCollectionBadge(lib.CollectionType) + '</td>';
         overviewHtml += '<td>' + formatBytes(lib.VideoSize) + '</td>';
         overviewHtml += '<td>' + formatBytes(lib.AudioSize) + '</td>';